requirements.txt
Flask==2.3.3
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-SQLAlchemy==3.1.1
email-validator==2.1.0.post1
python-dotenv==1.0.1
gunicorn==22.0.0
